+++ HISTORY.rst
@@ -8,6 +8,8 @@ Pending Release
 .. Modify the below with new release notes
 
 * Drop Python 2 support, only Python 3.4+ is supported now.
+* Add support for running `Black <https://pypi.org/project/black/>`__, the
+  Python code auto-formatter.
 
 2.4.0 (2018-09-30)
 ------------------